CMake & compile options

+++ commands/cat.hpp
@@ -36,7 +36,7 @@ public:
         }
 
         std::stringstream result;
-        int32_t error_count = 0;
+        size_t error_count = 0;
         for (auto &filename: params.args) {
             std::filesystem::path current_path(env->at("PWD").to_string());
             current_path /= filename;
